p9: remove sending FDs via Unix Domain Sockets
It was a really cool feature for gVisor, where all communication
happened on the same host. Cool optimization.

We intend to use this package to communicate across kernel boundaries,
and would like to be able to use TCP sockets rather than just Unix
domain sockets.

Hence, we remove the feature. It could make an optional come back some
day though.

Signed-off-by: Chris Koch <chrisko@google.com>