ensure that readdir only encodes an integral number of dirents
the 9p standard requires that readdir respond with an integral
number of dirents. The encoder for readdir could respond with
a partial dirent if adding the full dirent would exceed msize.
The Linux kernel in turn got upset and returned EIO to user mode.

This change fixes the problem and adds a test.

We need to start processing and honoring the offset in a readdir,
but probably *after* we fix Linux.

Should we be vendoring p9? dep ensure? What's the call here?

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>