Use existing `ulog`s rather than global std `log`
`Server` and `Client` have a default value for their log, which can be
overridden by callers of their constructors. If none are provided,
nothing should be printed, if one is, it should be used for these
messages, rather than using the global `log`.

`clientFile` does not need to log a failure from `Close` at all, the
caller will receive the error value directly, and the context added to
the message, is already implied by the protocol.