xattr: test case for 2 failures
Case 1: setting attr larger than msize

Documentation for Txattrcreate states that "The actual setxattr
operation happens when the fid is clunked. At that point the written
byte count and the attr_size specified in TXATTRCREATE should be same
otherwise an error will be returned."

I.e. writes should accumulate until Tclunk, when the actual operation
happens, so that they can be chunked over Msize.

Case 2: removexattr, which in 9P2000.L is a Txattrcreate(AttrSize: 0,
Flags: XATTR_REPLACE), which is unhandled currently.

Signed-off-by: Chris Koch <chrisko@google.com>