QID mapper: ensure QIDs are unique in composefs
composefs exposes several file systems as if they are one to the client.
QIDs have the property that they shall be unique on a server, with the
following properties according to intro(5):

    The qid represents the server’s unique identification for the file
    being accessed: two files on the same server hierarchy are the same
    if and only if their qids are the same. [...] The thirteen-byte qid
    fields hold a one-byte type, [...] and two unsigned integers: first
    the four-byte qid version, then the eight-byte qid path.

    The path is an integer unique among all files in the hierarchy. If a
    file is deleted and recreated with the same name in the same
    directory, the old and new path components of the qids should be
    different. The version is a version number for a file; typically, it
    is incremented every time the file is modified.

The QID mapper therefore ensures that the path integer is unique among
all files in its hierarchy, while keeping type and version the same.

Signed-off-by: Chris Koch <chrisko@google.com>