localfs: (WIP) `File.Readdir` 9P2000.L compliance
The count parameter for `Readdir` is a byte count (as in POSIX
`getdents`), not an integral count (as in Go).
This patch treats the value as such, and does estimates and conversions
from 9P -> Go to perform the operations within the expectations of the
9P standard.
Excess entries read during operation, are retained on the FID, to allow
subsequent calls to resume reading the directory with the appropriate
offset.
Rewinding the directory is also supported with offset `0`.
Works but is not considered stable yet.